[Fix] HC-112 불필요한 파일 코드 정리
