--- src/main/java/com/example/coconote/api/chatRoom/controller/ChatController.java
@@ -1,25 +1,0 @@
-//package com.example.coconote.api.chat.controller;
-//
-//import com.example.coconote.api.chat.entity.ChatRoom;
-//import com.example.coconote.api.chat.service.ChatService;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.web.bind.annotation.*;
-//
-//import java.util.List;
-//
-//@RequiredArgsConstructor
-//@RestController
-//@RequestMapping("api/v1/chat")
-//public class ChatController {
-//    private final ChatService chatService;
-//
-//    @PostMapping
-//    public ChatRoom createRoom(@RequestParam String name) {
-//        return chatService.createRoom(name);
-//    }
-//
-//    @GetMapping
-//    public List<ChatRoom> findAllRooms() {
-//        return chatService.findAllRoom();
-//    }
-//}
