--- src/main/java/com/example/coconote/api/chatRoom/dto/ChatMessage.java
@@ -1,18 +1,0 @@
-package com.example.coconote.api.chatRoom.dto;
-
-import com.example.coconote.api.chatRoom.entity.MessageType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ChatMessage {
-    private MessageType type;
-    private String roomId;
-    private String sender;
-    private String message;
-}
