--- src/main/java/com/example/coconote/api/chatRoom/entity/ChatRoom.java
@@ -1,41 +1,0 @@
-package com.example.coconote.api.chatRoom.entity;
-
-//import com.example.coconote.api.chat.service.ChatService;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-        import java.util.UUID;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ChatRoom {
-    private String roomId;
-    private String name;
-
-    public static ChatRoom create(String name){
-        ChatRoom chatRoom = new ChatRoom();
-        chatRoom.roomId = UUID.randomUUID().toString();
-        chatRoom.name = name;
-        return chatRoom;
-    }
-
-
-
-//    private Set<WebSocketSession> sessions = new HashSet<>();
-//
-//    public void handleActions(WebSocketSession session, ChatMessage chatMessage, ChatService chatService){
-//        if(chatMessage.getType().equals(MessageType.ENTER)){
-//            sessions.add(session);
-//            chatMessage.setMessage(chatMessage.getSender() + "님이 입장했습니다.");
-//        }
-//        sendMessage(chatMessage, chatService);
-//    }
-//
-//    public <T> void sendMessage(T message, ChatService chatService){
-//        sessions.parallelStream().forEach(session -> chatService.sendMessage(session, message));
-//    }
-}
