--- src/main/java/com/example/coconote/api/chatRoom/entity/MessageType.java
@@ -1,5 +1,0 @@
-package com.example.coconote.api.chatRoom.entity;
-
-public enum MessageType {
-    ENTER, TALK
-}
