--- src/main/java/com/example/coconote/api/chatRoom/repository/ChatRoomRepository.java
@@ -1,36 +1,0 @@
-package com.example.coconote.api.chatRoom.repository;
-
-import com.example.coconote.api.chatRoom.entity.ChatRoom;
-import jakarta.annotation.PostConstruct;
-import org.springframework.stereotype.Repository;
-
-import java.util.*;
-
-@Repository
-public class ChatRoomRepository {
-    private Map<String, ChatRoom> chatRoomMap;
-
-    @PostConstruct
-    public void init() {
-        chatRoomMap = new HashMap<>();
-    }
-
-    public List<ChatRoom> findAllRoom() {
-        List chatRooms = new ArrayList<>(chatRoomMap.values());
-        Collections.reverse(chatRooms);
-        return chatRooms;
-    }
-
-    public ChatRoom findRoomById(String roomId){
-        return chatRoomMap.get(roomId);
-    }
-
-    public ChatRoom createChatRoom(String name) {
-        ChatRoom chatRoom = ChatRoom.create(name);
-        chatRoomMap.put(chatRoom.getRoomId(), chatRoom);
-        return chatRoom;
-    }
-
-
-
-}
