--- src/main/java/com/example/coconote/api/chatRoom/service/ChatService.java
@@ -1,55 +1,0 @@
-//package com.example.coconote.api.chat.service;
-//
-//import com.example.coconote.api.chat.entity.ChatRoom;
-//import com.fasterxml.jackson.core.JsonProcessingException;
-//import com.fasterxml.jackson.databind.ObjectMapper;
-//import jakarta.annotation.PostConstruct;
-//import lombok.RequiredArgsConstructor;
-//import lombok.extern.slf4j.Slf4j;
-//import org.springframework.stereotype.Service;
-//import org.springframework.web.socket.TextMessage;
-//import org.springframework.web.socket.WebSocketSession;
-//
-//import java.io.IOException;
-//import java.util.*;
-//
-//@Slf4j
-//@RequiredArgsConstructor
-//@Service
-//public class ChatService {
-//
-//    private final ObjectMapper objectMapper;
-//    private Map<String, ChatRoom> chatRooms;
-//
-//    @PostConstruct
-//    public void init() {
-//        chatRooms = new LinkedHashMap<>();
-//    }
-//
-//    public List<ChatRoom> findAllRoom(){
-//        return new ArrayList<>(chatRooms.values());
-//    }
-//
-//    public ChatRoom findRoomById(String roomId){
-//        return chatRooms.get(roomId);
-//    }
-//
-//    public ChatRoom createRoom(String name){
-//        String randomId = UUID.randomUUID().toString();
-//        ChatRoom chatRoom = ChatRoom.builder()
-//                .roomId(randomId)
-//                .name(name)
-//                .sessions(new HashSet<>())
-//                .build();
-//        chatRooms.put(randomId, chatRoom);
-//        return chatRoom;
-//    }
-//
-//    public <T> void sendMessage(WebSocketSession session, T message) {
-//        try {
-//            session.sendMessage(new TextMessage(objectMapper.writeValueAsString(message)));
-//        } catch (IOException e) {
-//            log.error(e.getMessage(), e);
-//        }
-//    }
-//}
