+++ src/main/java/com/example/coconote/api/thread/controller/WebSocketController.java
@@ -2,12 +2,10 @@ package com.example.coconote.api.thread.controller;
 
 // import 생략...
 
-import com.example.coconote.api.thread.dto.requset.ChatMessage2;
-import com.example.coconote.api.chatRoom.entity.MessageType;
+import com.example.coconote.api.thread.entity.MessageType;
 import com.example.coconote.api.thread.dto.requset.ThreadCreateReqDto;
 import com.example.coconote.api.thread.dto.response.ThreadResDto;
 import com.example.coconote.api.thread.service.ThreadService;
-import com.example.coconote.api.thread.service.WebSocketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
