+++ src/main/java/com/example/coconote/api/thread/dto/requset/ChatMessage2.java
@@ -1,6 +1,6 @@
 package com.example.coconote.api.thread.dto.requset;
 
-import com.example.coconote.api.chatRoom.entity.MessageType;
+import com.example.coconote.api.thread.entity.MessageType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
