+++ src/main/java/com/example/coconote/api/thread/entity/MessageType.java
@@ -1,0 +1,5 @@
+package com.example.coconote.api.thread.entity;
+
+public enum MessageType {
+    ENTER, TALK
+}
